add analytics manager
